Crate zng_unique_id

This crate is part of the zng project.

Macros for generating unique ID types.

Re-exports

pub use zng_txt::Txt;

Macros

impl_unique_id_fmt

Implement debug and display for an unique ID type that also implements name.

impl_unique_id_name

Extend an unique ID type to have an optional attached name string, also implements.

unique_id_32

Declare a new unique id type that is backed by a NonZeroU32.

unique_id_64

Declare a new unique id type that is backed by a NonZeroU64.

Structs

BuildIdHasher

Build IdHasher.

IdHasher

No-op hasher.

IdMap

Map specialized for unique IDs that are already a randomized hash.

IdSet

Set specialized for unique IDs that are already a randomized hash.

Enums

IdNameError

Error when trying to associate give a name with an existing id.

Type Aliases

IdEntry

Entry in IdMap.

IdOccupiedEntry

Occupied entry in an IdEntry.

IdVacantEntry

Vacant entry in an IdEntry.